functions/align-kvectors/requirements-dev.txt
-r requirements.txt
python-dotenv==1.2.2
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.2
pytest-mock==3.15.1
